Fix Ansible syntax upgrade defun
It appended a space to the last converted attribute. This was unnoticed since
ws-butler automatically removed the trailing space, but was a problem for
attribute values that needed to be quoted (eg. `{{ item }}`, because in this
case the trailing space ended up within the quotation and was not removed by
ws-butler.