Fix logic to enable Unicode symbols in the modeline
I have two systemd-managed daemon instances of the Emacs daemon. One for the
terminal, and another for X. Keeping them separate allows for slightly different
configuration to avoid some issues that arise due to the difference between the
terminal and the graphical environment.

In the terminal I don't want Unicode symbols in the modeline because my terminal
emulator does not render them with the right spacing, which makes for a jarring
visual experience.

Previously the logic whether Unicode symbols in the modeline should be enabled
was to enable them when Emacs was not started by my terminal daemon service.
This caused it to choose incorrectly when Emacs was started on the terminal
without the use of the daemon (`emacs -nw`).

This change adjusts the logic to enable the Unicode symbols when Emacs either
has a graphical backend (eg. X) or was started by my X daemon script. This
logic, I think, will choose correctly in all cases.