Enable transient paste mode
The develop branch of Spacemacs changed the key bindings for next / prev to
`C-n` / `C-p`, and continuously pressing `p` repeats the paste operation.

This is much saner then the key bindings on the master branch, which were `n` /
`p` and no way to repeat the operation.

This feature is usable for me now.