Remove customization of `spacemacs/sort-lines`
The defun on the Spacemacs develop branch now uses the prefix argument itself,
to trigger reverse sorting, that conflicts with our customization.

This removes the customization and substitutes it with a toggle for
`sort-fold-case`. This is a better way to go about case insensitive sorting
anyway, because the flag influences many more sorting functions.