Revise go linting settings
This restricts gometalinter to run the fast linters only - the slow ones took up
to 10 seconds, which is just not an acceptable time frame.

The default go build command is invoked with the `-i` flag now. This is required
to make the auto-completion with `gocode` work reliably for third party library
dependencies.