Minor adjustments to `xclip-mode`
Switch back to using `xclip` from `xsel`. Both offer practically the same
functionality anyway, and according to the principle of least surprise, it is
better if `xclip-mode` actually uses `xclip`.

The logic to automatically activate the mode is changed as well, so it
activatesonly in the TTY Emacs daemon, and not in the Emacs daemon that's
indented for graphical environments.