Remove the org Trello integration
The `org-trello` package does not play nice with the autoload mechanism. It
automatically sets a customize variable when the package loads, which
immediately gets persisted into `custom.el`. The next time Emacs starts, this
variable triggers an automatic require of the `org-trello` package.

The package is just not worth dealing with this. Gone it is.