Many many improvements to the Ansible layer
I'd like to put a detailed explanation here, but truth is, this is the work
of an entire year and I've forgotten most of it already. So here is a summary
commit at least.

The only thing I know of that is not properly working right now is font-lock.
Argument names are still font-locked as commands if they happen to have the same
name as a command. This can be fixed, but not today.