Remove unbound package, it breaks the Spacemacs package updater
The error message is:

    version-to-list: Invalid version syntax: ‘Wrong type argument:
    number-or-marker-p, nil’ (must start with a number)

It probably has something to do with how the package metadata at the top of
unbound.el is formatted. I ran the debugger on the error and tried to understand
what's going on so I may eventually find a workaround or fix, but I did get
nowhere.

The package is not worth the hassle, so I am going to remove it.