Rename the 'mfa-' namespace to 'my-'
Initially I chose 'mfa' as name of the namespace for my own additions because
they are my initials, and seemed unique enough. But that choice has often felt
wrong. It's unnecessary specific, maybe a little bit vain.

Well, now is as good a time as any to do a mass rename to a more neutral name.