Clean up customization of org mode faces
This did not need to be in a pre-init hook, a regular post-init does suffice.

The function to customize the faces not only needs to be hooked, it also needs
to be called immediately. The hook is only called after a theme is loaded, and
the theme may have already been loaded at this point.