Make the dired customization mess a little bit less messy
The trigger to clean this up was to fix an error when trying to load a saved
Spacemacs layout that contained a dired buffer. The problem was the order of
autoloaded packages, right after startup the treemacs icons might not be loaded
yet, which causes an error when treemacs-icons-dired tries to insert a "nil"
icon into the dired buffer.

While I was already changing things, this also removes the obsolete
all-the-icons-dired customizations, simplifies the treemacs integration with
dired-subtree, and includes some minor inconsequential fixes.