A basic "(auto save / resume) project layout" functionality
Usage is the same as "helm-projectile-switch-project", but this function will
automatically create a perspective with the same name as the project and
configure it to auto-save the perspective to a file named ".layout.el" in the
project root directory.

Any subsequent use of the function will resume the saved perspective.