Fix YAML font lock glitches and improve Ansible font lock rules
This fixes the font lock glitches with YAML block literals by supplying
syntax-propertize-extend-region and font-lock-extend-region functions.

This also includes adjustments to the Ansible font-lock rules, the new rules
attempt to make the highlighting semantically aware, eg. the YAML key "user"
will be fontified as a keyword if it is used as a task name, not when it is an
argument to the "file" task. These new font-lock rules are an experiment, they
are heuristics and thus not always accurate, and they might be expensive in
terms of CPU time. If they pose a problem, I may remove them later.