Only enable dtrt-indent mode in buffers visiting a file
Initially the idea of excluding the *scratch* buffer was to avoid dtrt-indent.el
to load eagerly at startup, which it would because my '*scratch*' buffer uses
text-mode.

The logic was always iffy because it relies on the name of the scratch buffer,
which de facto always is *scratch*, but there is no real guarantee for that.

After thinking about it, let's exclude all buffers that are not visiting a file.
That avoid having to match against the name of the scratch buffer, and will also
exclude some other buffers that dtrt-indent mode is useless for.