Fix return to magit status buffer after quitting commit buffer
Magit and with-editor try to be helpful and restore the previous window config
after a server buffer quits. Thing is, if we alter the window config in
server-after-make-frame-hook, the altered window config will get restored, which
in this case means the git commit buffer always return to the Spacemacs home
buffer. Not ok.

This hook was intended to fix the issue that, for an unknown reason, sometimes
when the emacsclient opens a new frame that frame already has a vertical or
horizontal window split. I'll remove the hook for now; if the issue re-occurs,
I'll try to find a different fix that hopefully does not break magit workflows.