Remove attempts to alter the evil state during yasnippet field navigation
Things do seem to work with evil out of the box now, there are no errors if evil
visual or insert state is active when snippet expansion is triggered.

This attempt to enter insert state when navigating to a field was always a bit
buggy and had potential for unintended side effects. Let's just give up trying
to enter evil insert state automatically, it is too much hassle to make it work
robustly.